requirements.txt
Flask
Flask-CORS
Flask-JWT-Extended
psycopg2-binary
python-dotenv
bcrypt
resend
qrcode
Pillow
gunicorn
